pyproject.toml
[tool.poetry]
name = "shiro"
version = "0.1.0"
description = "A Discord bot for chess players"
authors = ["eniraa <47072520+eniraa@users.noreply.github.com>"]
license = "LGPL-3.0-only"

[tool.poetry.dependencies]
python = "3.10"
hikari = {extras = ["speedups"], version = "^2.0.0.dev104"}
hikari-lightbulb = "^2.0.1"
uvloop = {version = "^0.16.0", markers = "sys_platform != 'win32'"}
python-dotenv = "^0.19.1"
requests = "^2.26.0"
PyYAML = "^6.0"
beautifulsoup4 = "^4.10.0"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
flake8-annotations = "^2.7.0"
flake8-bandit = "^2.1.2"
flake8-docstrings = "^1.6.0"
flake8-isort = "^4.1.1"
isort = "^5.10.1"
pre-commit = "^2.15.0"
taskipy = "^1.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
reformat = {cmd = "black . && isort .", help = "Reformats code to conform to some lints"}
black = {cmd = "black --check .", help = "Dry run of black"}
flake8 = {cmd = "python -m flake8", help = "Lints code with flake8"}
lint = {cmd = "pre-commit run --all-files", help = "Checks all files for CI errors"}
precommit = {cmd = "pre-commit install --install-hooks", help = "Installs the precommit hook"}
